requirements.txt
asgiref==3.9.1
django==5.2.5
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-rest-framework==0.1.0
django-unfold==0.63.0
djangorestframework==3.16.1
psycopg==3.2.3
psycopg-binary==3.2.3
sqlparse==0.5.3
typing-extensions==4.14.1
tzdata==2025.2
